Introduce vdj_types crate
