+++ vdj_ann/Cargo.toml
@@ -30,3 +30,4 @@ string_utils = "0.1.1"
 vector_utils = "0.1.3"
 strum = ">=0.17.1"
 strum_macros = ">=0.17.1"
+vdj_types = { path = "../vdj_types" }
