+++ vdj_types/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "vdj_types"
+version = "0.1.0"
+authors = ["Sreenath Krishnan <sreenath.krishnan@10xgenomics.com>"]
+edition = "2018"
+license = "MIT"
+
+
+[dependencies]
+serde = { version = "1.0.90", features = ["derive"] }
+serde_json = "1.0.41"
+strum = ">=0.17.1"
+strum_macros = ">=0.17.1"
